refactor: move query invitations to query-flows.contract.js
- the query invitations that post to vstorage are not representative of an example contract
- move these invitations to a new test/fixtures/query-flows.contract.js so basic-flows.contact.js
  can stand on its own as an example
