+++ packages/orchestration/src/examples/basic-flows.contract.js
@@ -40,10 +40,6 @@ const contract = async (
     M.interface('Basic Flows PF', {
       makeOrchAccountInvitation: M.callWhen().returns(InvitationShape),
       makePortfolioAccountInvitation: M.callWhen().returns(InvitationShape),
-      makeSendICQQueryInvitation: M.callWhen().returns(InvitationShape),
-      makeAccountAndSendBalanceQueryInvitation:
-        M.callWhen().returns(InvitationShape),
-      makeSendLocalQueryInvitation: M.callWhen().returns(InvitationShape),
     }),
     {
       makeOrchAccountInvitation() {
@@ -56,24 +52,6 @@ const contract = async (
         return zcf.makeInvitation(
           orchFns.makePortfolioAccount,
           'Make an Orchestration Account',
-        );
-      },
-      makeSendICQQueryInvitation() {
-        return zcf.makeInvitation(
-          orchFns.sendICQQuery,
-          'Submit a query to a remote chain',
-        );
-      },
-      makeAccountAndSendBalanceQueryInvitation() {
-        return zcf.makeInvitation(
-          orchFns.makeAccountAndSendBalanceQuery,
-          'Make an account and submit a balance query',
-        );
-      },
-      makeSendLocalQueryInvitation() {
-        return zcf.makeInvitation(
-          orchFns.sendLocalQuery,
-          'Submit a query to the local chain',
         );
       },
     },
