+++ packages/orchestration/src/fixtures/README.md
@@ -1,0 +1,3 @@
+# Fixtures
+
+This directory contains contracts used for testing. They are NOT representative of patterns we'd like to use in production, and only exist to facilitate testing.
